Grouped codes related to GUI into a package `com.gui`.
