+++ src/main/java/com/gui/CalculatorGUI.java
@@ -1,18 +1,5 @@
-/*-------------------------------------*/
-
-/* README */
-
-// # Structjr
-
-//TODO: Currently, `CalculatorGUI.cart` and `CalculatorGUI.subtotalEntryMap` are synchronized manually though they are semantically linked strongly.
-
-/*-------------------------------------*/
-
-/* import */
-
 package com.gui;
 
-import java.io.File;
 import java.util.List;
 import java.util.Map;
 import java.util.HashMap;
@@ -23,14 +10,9 @@ import javax.swing.*;
 import javax.imageio.*;
 
 import com.example.Calculator;
-
-/*-------------------------------------*/
-
-/* CalculatorGUI */
+import com.gui.CalculatorMock;
 
 public class CalculatorGUI extends JFrame {
-
-    /* fields */
 
     //GUI {
 
@@ -45,13 +27,14 @@ public class CalculatorGUI extends JFrame {
     private static final int    windowWidth  = 1000;
     private static final int    windowHeight = 400;
 
-    public static Color defaultBgColor = Color.gray;
+    public static final Color defaultBgColor = Color.gray;
 
     //} GUI
 
     //CUI {
 
-    private static Calculator calculator;
+//     private static Calculator calculator;
+    private static CalculatorMock calculator;
 
     public static Map<String, Integer> cart;
 
@@ -68,7 +51,9 @@ public class CalculatorGUI extends JFrame {
         //CUI
         {
 
-            CalculatorGUI.calculator = new Calculator();
+//             CalculatorGUI.calculator = new Calculator();
+            CalculatorGUI.calculator = new CalculatorMock();
+
             CalculatorGUI.cart = new HashMap<>();
             CalculatorGUI.subtotalEntryMap = new HashMap<>();
 
@@ -196,260 +181,4 @@ public class CalculatorGUI extends JFrame {
     }
 
 }
-
-/*-------------------------------------*/
-
-/* SubtotalEntry */
-
-class SubtotalEntry {
-
-    private static final int labelGridWidth = 5;
-    private static final int labelWidth = 16;
-
-    public JPanel     panel;
-    public JLabel     icon;
-    public JButton    resetButton;
-    public JButton    decrementButton;
-    public JButton    incrementButton;
-    public JTextField subtotalLabel;
-
-    public void setSubtotalLabelText(String productName, int quantity) {
-        this.subtotalLabel.setText( String.format("  (%d点)    %s", quantity, productName) );
-    }
-
-    public SubtotalEntry(String productName) {
-
-        panel = new JPanel();
-        panel.setBackground(Color.lightGray);
-        panel.setBorder(BorderFactory.createLineBorder(CalculatorGUI.defaultBgColor));
-        panel.setLayout(new GridBagLayout());
-        GridBagConstraints gridder = new GridBagConstraints();
-
-        int gridCounter = 0;
-
-        decrementButton = new JButton("-");
-        decrementButton.addActionListener(new IncrementButtonHandler(productName, /* numIncrement = */ -1));
-        gridder.fill = GridBagConstraints.HORIZONTAL;
-        gridder.gridx = gridCounter++;
-        gridder.gridy = 0;
-        gridder.gridwidth = 1;
-        panel.add(this.decrementButton, gridder);
-
-        incrementButton = new JButton("+");
-        incrementButton.addActionListener(new IncrementButtonHandler(productName));
-        gridder.fill = GridBagConstraints.HORIZONTAL;
-        gridder.gridx = gridCounter++;
-        gridder.gridy = 0;
-        gridder.gridwidth = 1;
-        panel.add(this.incrementButton, gridder);
-
-        try {
-            icon = new JLabel(IconGenerator.generateMiniIcon(productName));
-        } catch (Exception e) {
-            ;
-        }
-        gridder.fill = GridBagConstraints.HORIZONTAL;
-        gridder.gridx = gridCounter++;
-        gridder.gridy = 0;
-        gridder.gridwidth = 1;
-        panel.add(this.icon, gridder);
-
-        subtotalLabel = new JTextField(SubtotalEntry.labelWidth);
-        subtotalLabel.setEditable(false);
-        subtotalLabel.setBackground(Color.lightGray);
-        subtotalLabel.setBorder(null);
-        setSubtotalLabelText(productName, 1);
-        gridder.gridx = gridCounter;
-        gridCounter += SubtotalEntry.labelGridWidth;
-        gridder.gridy = 0;
-        gridder.gridwidth = SubtotalEntry.labelGridWidth;
-        panel.add(this.subtotalLabel, gridder);
-
-        resetButton = null;
-        try {
-            resetButton = new JButton(new ImageIcon(ImageIO.read(new File("./icons/reset_button_2.png"))));
-        } catch (Exception e) {
-            ;
-        }
-        resetButton.setBackground(Color.lightGray);
-        resetButton.setBorder(null);
-        resetButton.addActionListener(new ResetButtonHandler(productName));
-
-        gridder.gridx = gridCounter++;
-        gridder.gridy = 0;
-        gridder.gridwidth = 1;
-        panel.add(this.resetButton, gridder);
-
-    }
-
-}
-
-/*-------------------------------------*/
-
-/* IconGenerator */
-
-class IconGenerator {
-
-    private static final String iconDirectory = "./icons/";
-
-    private static Map<String, String> iconPathMap; //maps a product name to its icon path
-    private static Map<String, String> miniIconPathMap; //ditto but with smaller icons
-
-    static {
-
-        IconGenerator.iconPathMap     = new HashMap<>();
-        IconGenerator.miniIconPathMap = new HashMap<>();
-
-        iconPathMap.put("りんご",           "apple.png");
-        iconPathMap.put("みかん",           "orange.png");
-        iconPathMap.put("ぶどう",           "grapes.png");
-        iconPathMap.put("のり弁",           "nori_ben.png");
-        iconPathMap.put("しゃけ弁",         "sake_ben.png");
-        iconPathMap.put("タバコ",           "cigarette.png");
-        iconPathMap.put("メンソールタバコ", "menthol_cigarette.png");
-        iconPathMap.put("ライター",         "lighter.png");
-        iconPathMap.put("お茶",             "tea.png");
-        iconPathMap.put("コーヒー",         "coffee.png");
-        iconPathMap.put("光のハンバーガー", "hamburger.png");
-
-        iconPathMap.put("りせっと",         "all_reset_2.png");
-
-        miniIconPathMap.put("りんご",           "apple_22.png");
-        miniIconPathMap.put("みかん",           "orange_22.png");
-        miniIconPathMap.put("ぶどう",           "grapes_22.png");
-        miniIconPathMap.put("のり弁",           "nori_ben_22.png");
-        miniIconPathMap.put("しゃけ弁",         "sake_ben_22.png");
-        miniIconPathMap.put("タバコ",           "cigarette_22.png");
-        miniIconPathMap.put("メンソールタバコ", "menthol_cigarette_22.png");
-        miniIconPathMap.put("ライター",         "lighter_22.png");
-        miniIconPathMap.put("お茶",             "tea_22.png");
-        miniIconPathMap.put("コーヒー",         "coffee_22.png");
-        miniIconPathMap.put("光のハンバーガー", "hamburger_22.png");
-
-    }
-
-    private static ImageIcon __generateIcon(String productName, Map<String, String> iconPathMap) {
-        ImageIcon ret = null;
-        try {
-            final String iconPath = IconGenerator.iconDirectory + iconPathMap.get(productName);
-            ret = new ImageIcon(ImageIO.read(new File(iconPath)));
-        } catch (Exception e) {
-            ;
-        }
-        return ret;
-    }
-
-    public static ImageIcon generateIcon(String productName) {
-        return IconGenerator.__generateIcon(productName, IconGenerator.iconPathMap);
-    }
-
-    public static ImageIcon generateMiniIcon(String productName) {
-        return IconGenerator.__generateIcon(productName, IconGenerator.miniIconPathMap);
-    }
-
-}
-
-/*-------------------------------------*/
-
-/* AllResetButtonHandler */
-
-class AllResetButtonHandler implements ActionListener {
-
-    public void actionPerformed(ActionEvent event) {
-        CalculatorGUI.removeAllSubtotalEntry();
-    }
-
-}
-
-/*-------------------------------------*/
-
-/* ResetButtonHandler */
-
-class ResetButtonHandler implements ActionListener {
-
-    private String productName;
-
-    public ResetButtonHandler (String productName) {
-        this.productName = productName;
-    }
-
-    public void actionPerformed(ActionEvent event) {
-        CalculatorGUI.removeSubtotalEntry(this.productName);
-    }
-
-}
-
-/*-------------------------------------*/
-
-/* IncrementButtonHandler */
-
-class IncrementButtonHandler implements ActionListener {
-
-    private String productName;
-    private int numIncrement;
-
-    public IncrementButtonHandler(String productName) {
-        this(productName, 1);
-    }
-
-    public IncrementButtonHandler(String productName, int numIncrement) {
-        this.productName = productName;
-        this.numIncrement = numIncrement;
-    }
-
-    public void actionPerformed(ActionEvent event) {
-        int quantity = CalculatorGUI.cart.get(this.productName);
-        if (quantity + this.numIncrement <= 0) { //if in decrement mode and the current quantity is 1
-            CalculatorGUI.removeSubtotalEntry(this.productName);
-        } else {
-            quantity += this.numIncrement;
-            CalculatorGUI.cart.put(this.productName, quantity);
-            CalculatorGUI.subtotalEntryMap.get(productName).setSubtotalLabelText(productName, quantity);
-            CalculatorGUI.recalculateTotalAmountDisplay();
-            CalculatorGUI.redraw();
-        }
-    }
-
-}
-
-/*-------------------------------------*/
-
-/* ProductButtonHandler */
-
-class ProductButtonHandler implements ActionListener {
-
-    public void actionPerformed(ActionEvent event) {
-
-        //ex: "<html>りんご<br/>(100円)</html>" -> "りんご"
-        String productName = event.getActionCommand();
-        var pattern = java.util.regex.Pattern.compile("(?<=<html>)[^<]+");
-        var matcher = pattern.matcher(productName);
-        matcher.find();
-        productName = matcher.group();
-
-        if (CalculatorGUI.cart.containsKey(productName)) { //increments the quantity
-
-            final int quantity = CalculatorGUI.cart.get(productName) + 1;
-            CalculatorGUI.cart.put(productName, quantity);
-            CalculatorGUI.subtotalEntryMap.get(productName).setSubtotalLabelText(productName, quantity);
-
-        } else { //creates an entry
-
-            CalculatorGUI.cart.put(productName, 1);
-
-            final SubtotalEntry entry = new SubtotalEntry(productName);
-            CalculatorGUI.panel2.add(entry.panel);
-            CalculatorGUI.subtotalEntryMap.put(productName, entry);
-
-        }
-        
-        CalculatorGUI.recalculateTotalAmountDisplay();
-
-        CalculatorGUI.redraw();
-
-    }
-
-}
-
-/*-------------------------------------*/
